Firebase
Added hooks into Firebase.

https://console.firebase.google.com/project/csee5590lab6/database/csee5590lab6/data/groceries

+++ app/src/main/java/com/jarrm5/sqlitespeechtotextapp/GroceryItemAdapter.java
@@ -67,7 +67,7 @@ public class GroceryItemAdapter extends ArrayAdapter<GroceryItem>  {
         mItemEditCircle.setColor(ContextCompat.getColor(getContext(),R.color.edit_background));
         mItemDeleteCircle.setColor(ContextCompat.getColor(getContext(),R.color.edit_background));
 
-        mItemQuantity.setText(Integer.toString(current.getQuantity()));
+        mItemQuantity.setText(Long.toString(current.getQuantity()));
         mItemName.setText(current.getName());
 
         mItemDelete.setOnClickListener(new View.OnClickListener() {
